feat: add delete and edit functionality to calendar content management
- Implement delete and edit modals for scheduled content in CalendarComponent
- Enhance ContentModal to handle delete and edit actions via props
- Refactor ContentEditModal and DeleteConfirmationModal for improved user experience
- Integrate toast notifications for successful content updates and deletions
